requirements.txt
# ---------------- Core ----------------
Flask==2.2.5
python-dotenv==1.0.1
itsdangerous==2.1.2

# ---------------- Database ----------------
mysql-connector-python==9.0.0    # For Koha DB (MySQL/MariaDB)
sqlite-utils==3.35
SQLAlchemy==2.0.36               # ORM (optional, future use)

# ---------------- Data Handling ----------------
# pandas 2.2.3 has official Windows wheels for Python 3.13 (unlike 2.2.2)
pandas==2.2.3
openpyxl==3.1.5
XlsxWriter==3.2.0
odfpy==1.4.1                     # For writing .ods files

# ---------------- PDF Generation ----------------
reportlab==4.2.2
Pillow==10.4.0                   # For image formats (PNG, etc.)

# RTL shaping for Arabic (so text joins correctly)
arabic-reshaper>=3.0.0,<4.0.0
python-bidi==0.4.2

# WeasyPrint requires system-level dependencies (GTK, Cairo, Pango).
# Optional on Windows; uncomment only if you have these installed.
# weasyprint==61.2 ; sys_platform != "win32"

# ---------------- Email ----------------
email-validator==2.2.0
Flask-Mail==0.9.1

# ---------------- Scheduling ----------------
APScheduler==3.10.4
Flask-APScheduler==1.13.1

# ---------------- Authentication & Security ----------------
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==2.3.8

# ---------------- Misc Utilities ----------------
requests==2.32.3
python-dateutil==2.9.0.post0
